delete PhotoNavigationController
